chore(tests): fix linter
